feat: add option for omit function key/value in the debugger
Basically, there were situations where it was difficult to grasp the values at a glance in the
debugger due to the functions being displayed, so we added an option to remove them. This is set to
true by default, and in the backup, the functions will always be excluded from the backup according
to the original logic. Please let me know your thoughts on this.